Working multi-depth
Both the char-generator and the text-spewer now handle a range of
possible depth values in the character model. I allow up to 9, but
in practise this is unlikely to be useful. The default, 3, is probably
the sanest setting, with 2/4 being an interesting version to compare.

Any higher, and you start over-constraining the model (plus, it takes ages
to build it and the file gets really big). But the option is there.

The text-spewer uses whatever value of depth was encoded in the presented
model.